logging: factor Data class out of MRA-tickbin for use elsewhere
