+++ MRA-tickbin.py
@@ -30,39 +30,13 @@ import sys
 import os
 os.environ["PROTOCOL_BUFFERS_PYTHON_IMPLEMENTATION"] = "python"
 import argparse
-import json
-import copy
-import struct
-import importlib
 from google.protobuf.message import Message
 from google.protobuf.descriptor import FieldDescriptor
 from google.protobuf.json_format import MessageToJson
 
-# get common datatypes generated by protobuf, take cmake build/ folder
-# it works for now, but TODO: make this nicer ...
-try:
-    sys.path.append(os.path.join(os.path.dirname(__file__), 'build'))
-    import datatypes.Meta_pb2
-    # the type of the header is always the same
-    # based on the content of the header, dynamic imports can be done
-except ImportError:
-    raise ImportError('failed to import datatypes.Meta_pb2, perhaps cmake build was not done?')
-    # or try bazel build location?
+# MRA tickbin Data class
+from libraries.logging.logdata import Data, FILE_ELEMENTS
 
-
-
-# constants
-FILE_ELEMENTS = (
-    'meta',
-    'input',
-    'params',
-    'state_before',
-    'output',
-    'local',
-    'state_after',
-    )
-EXPECTED_FILE_ELEMENT_COUNT = len(FILE_ELEMENTS)
-# these are all protobuf objects
 
 
 def parse_args(args: list) -> argparse.Namespace:
@@ -80,54 +54,6 @@ def parse_args(args: list) -> argparse.Namespace:
     parser.add_argument('-s', '--state', help='also print state, both before and after, instead of hiding', action='store_true')
     parser.add_argument('datafile', help='data file (.bin) to load')
     return parser.parse_args(args)
-
-
-
-class Data():
-
-    def __init__(self, filename):
-        self.load(filename)
-
-    def load(self, filename):
-        # load all binary data, uninterpreted
-        self.raw_load(filename)
-        # parse meta header to find out the type(s) to use when interpreting raw data
-        self.parse_meta_header()
-        # assign raw_data to types of the current component
-        self.parse_component_data()
-
-    def raw_load(self, filename):
-        self.raw_data = {}
-        loaded_elem_count = 0
-        with open(filename, 'rb') as f:
-            for key in FILE_ELEMENTS:
-                n = struct.unpack('i', f.read(4))[0]
-                self.raw_data[key] = f.read(n)
-                loaded_elem_count += 1
-        if loaded_elem_count != EXPECTED_FILE_ELEMENT_COUNT:
-            print('WARNING: file has {:d} elements, but {:d} were expected'.format(loaded_elem_count, EXPECTED_FILE_ELEMENT_COUNT))
-
-    def parse_meta_header(self):
-        self.meta = datatypes.Meta_pb2.Meta()
-        self.meta.MergeFromString(self.raw_data['meta'])
-
-    def parse_component_data(self):
-        for elem in FILE_ELEMENTS[1:]: # skip first element (meta)
-            key = elem.split('_')[0]
-            Key = key.capitalize() # example Input
-            # dynamic import of component interface
-            key_module = '{:s}_pb2'.format(Key) # example Input_pb2
-            module_name = 'components.{:s}.interface.{:s}'.format(self.meta.subfolder.replace('/', '.'), key_module, Key)
-            try:
-                module = importlib.import_module(module_name)
-            except:
-                continue
-            # construct message and merge from raw data
-            msg = getattr(module, Key)()
-            msg.MergeFromString(self.raw_data[elem])
-            # store
-            setattr(self, elem, msg)
-
 
 
 def protobuf_message_reduction(message: Message, options: dict = {}) -> Message:
